requirements.txt
streamlit
streamlit_calendar
pathlib
python-telegram-bot==20.8
